requirements.txt
fastapi==0.104.1
psycopg2==2.9.9
pyTelegramBotAPI==4.14.0
sqlalchemy==2.0.23
python-jose==3.3.0
PyJWT==2.8.0
bcrypt==4.1.1
fastapi_login==1.9.2